refactor: clear the codes of plugin jenkins-pipeline-kubernetes
