+++ internal/pkg/plugin/jenkinspipelinekubernetes/update.go
@@ -14,7 +14,7 @@ func Update(options map[string]interface{}) (map[string]interface{}, error) {
 		return nil, err
 	}
 
-	if errs := validateAndHandleOptions(&opts); len(errs) != 0 {
+	if errs := ValidateAndDefaults(&opts); len(errs) != 0 {
 		for _, e := range errs {
 			log.Errorf("Options error: %s.", e)
 		}
